+++ client/js/rl/test.html
@@ -30,6 +30,6 @@
     </div>
   </section>
  <!----- /main ----->
- <script src="/public-rl/test.bundle.js" type="module"></script>
+ <script src="/rl/public/test.bundle.js" type="module"></script>
  </body>
 </html>
