+++ client/js/rl/train.html
@@ -20,6 +20,6 @@
     </div>
   </section>
  <!----- /main ----->
- <script src="/public-rl/train.bundle.js" type="module"></script>
+ <script src="/rl/public/train.bundle.js" type="module"></script>
  </body>
 </html>
